+++ helpers/lib/index.js
@@ -22,7 +22,7 @@ const errorFormatters = require('./error-formatters')
 const httpFormatters = require('./http-formatters')
 
 module.exports = {
-  version: '1.5.0',
+  version: '1.6.0',
   stringify,
   ...errorFormatters,
   ...httpFormatters