+++ helpers/test/hapi.test.js
@@ -59,7 +59,7 @@ test('hapi res/req serialization', testOpts, t => {
     t.deepEqual(rec.http, {
       version: '1.1',
       request: {
-        method: 'get',
+        method: 'GET',
         headers: {
           host: `localhost:${server.info.port}`,
           connection: 'close'